separate test_output.py and test_unit.py; try to use require_ds decorator
